Last pass (hopefully) of refactoring for outside venus

+++ mars/tools/BHTSimulator.java
@@ -185,8 +185,8 @@ public class BHTSimulator extends AbstractMarsToolAndApplication implements Acti
         m_gui.getTfAddress().setText("");
         m_gui.getTfIndex().setText("");
         m_gui.getTaLog().setText("");
-        m_bhtModel.initBHT(((Integer) m_gui.getCbBHTentries().getSelectedItem()).intValue(),
-                ((Integer) m_gui.getCbBHThistory().getSelectedItem()).intValue(),
+        m_bhtModel.initBHT((Integer) m_gui.getCbBHTentries().getSelectedItem(),
+                (Integer) m_gui.getCbBHThistory().getSelectedItem(),
                 ((String) m_gui.getCbBHTinitVal().getSelectedItem()).equals(BHTSimGUI.BHT_TAKE_BRANCH));
 
         m_pendingBranchInstAddress = 0;
